Delete LICENSE
Purely free cannot prosper. This social condition needs to run things as a campaign, even if it is a diamond in the rough. 
And, also, I see Apple and Google are not protecting users' privacy as they could, according to official sources. We should save lives first as always, but we should never abuse power as we could avoid.